Rename bindGraphQLSchemaAPIToContext to bindG and include functions that dont need to be bound to context
