+++ .changeset/pretty-knives-rule.md
@@ -1,0 +1,5 @@
+---
+"@graphql-ts/schema": major
+---
+
+`bindGraphQLSchemaAPIToContext` and `GraphQLSchemaAPIWithContext` have been replaced with `initG` and `GWithContext` which also include all the APIs that aren't specifically bound to a context.
