+++ packages/schema/src/api-with-context/api-with-context.js
@@ -1,9 +1,27 @@
-import { bindGraphQLSchemaAPIToContext } from "../output";
+import { initG } from "../output";
 
-const __graphql = bindGraphQLSchemaAPIToContext();
+const __graphql = initG();
 
-export const { field, fields, interfaceField, object, union } = __graphql;
+export const {
+  field,
+  fields,
+  interfaceField,
+  object,
+  union,
+  Boolean,
+  Float,
+  ID,
+  Int,
+  String,
+  arg,
+  enumValues,
+  inputObject,
+  list,
+  nonNull,
+  scalar,
+} = __graphql;
 
 const interfaceType = __graphql.interface;
+const enumType = __graphql.enum;
 
-export { interfaceType as interface };
+export { interfaceType as interface, enumType as enum };
