+++ packages/schema/src/api-with-context/index.ts
@@ -1,5 +1,3 @@
-// this is written like this because an export * from './something'
-// where ./something uses `export = ` is not allowed
 export {
   field,
   fields,
